+++ workers/unity/Packages/io.improbable.gdk.gameobjectcreation/GameObjectInitializationSystem.cs
@@ -1,4 +1,4 @@
-using System.Linq;
+using System;
 using Improbable.Gdk.Core;
 using Improbable.Gdk.Subscriptions;
 using Unity.Entities;
@@ -26,23 +26,12 @@ namespace Improbable.Gdk.GameObjectCreation
         private EntityQuery newEntitiesQuery;
         private EntityQuery removedEntitiesQuery;
 
-        private ComponentType[] minimumComponentSet = new[]
-        {
-            ComponentType.ReadOnly<SpatialEntityId>()
-        };
+        private readonly EntityTypeExpectations entityTypeExpectations = new EntityTypeExpectations();
 
         public GameObjectInitializationSystem(IEntityGameObjectCreator gameObjectCreator, GameObject workerGameObject)
         {
             this.gameObjectCreator = gameObjectCreator;
             this.workerGameObject = workerGameObject;
-
-            var minCreatorComponentSet = gameObjectCreator.MinimumComponentTypes;
-            if (minCreatorComponentSet != null)
-            {
-                minimumComponentSet = minimumComponentSet
-                    .Concat(minCreatorComponentSet)
-                    .ToArray();
-            }
         }
 
         protected override void OnCreate()
@@ -58,6 +47,11 @@ namespace Improbable.Gdk.GameObjectCreation
                 Linker.LinkGameObjectToSpatialOSEntity(new EntityId(0), workerGameObject);
             }
 
+            var minimumComponentSet = new[]
+            {
+                ComponentType.ReadOnly<SpatialEntityId>(), ComponentType.ReadOnly<Metadata.Component>()
+            };
+
             newEntitiesQuery = GetEntityQuery(new EntityQueryDesc()
             {
                 All = minimumComponentSet,
@@ -69,6 +63,8 @@ namespace Improbable.Gdk.GameObjectCreation
                 All = new[] { ComponentType.ReadOnly<GameObjectInitSystemStateComponent>() },
                 None = minimumComponentSet
             });
+
+            gameObjectCreator.PopulateEntityTypeExpectations(entityTypeExpectations);
         }
 
         protected override void OnDestroy()
@@ -87,24 +83,68 @@ namespace Improbable.Gdk.GameObjectCreation
 
         protected override void OnUpdate()
         {
-            Entities.With(newEntitiesQuery).ForEach((Entity entity, ref SpatialEntityId spatialEntityId) =>
+            if (!newEntitiesQuery.IsEmptyIgnoreFilter)
             {
-                gameObjectCreator.OnEntityCreated(new SpatialOSEntity(entity, EntityManager), Linker);
-                PostUpdateCommands.AddComponent(entity, new GameObjectInitSystemStateComponent
-                {
-                    EntityId = spatialEntityId.EntityId
-                });
-            });
+                ProcessNewEntities();
+            }
 
+            if (!removedEntitiesQuery.IsEmptyIgnoreFilter)
+            {
+                ProcessRemovedEntities();
+            }
+
+            Linker.FlushCommandBuffer();
+        }
+
+        private void ProcessRemovedEntities()
+        {
             Entities.With(removedEntitiesQuery).ForEach((ref GameObjectInitSystemStateComponent state) =>
             {
                 Linker.UnlinkAllGameObjectsFromEntityId(state.EntityId);
-                gameObjectCreator.OnEntityRemoved(state.EntityId);
+
+                try
+                {
+                    gameObjectCreator.OnEntityRemoved(state.EntityId);
+                }
+                catch (Exception e)
+                {
+                    Debug.LogException(e);
+                }
             });
 
-            Linker.FlushCommandBuffer();
-
             EntityManager.RemoveComponent<GameObjectInitSystemStateComponent>(removedEntitiesQuery);
+        }
+
+        private void ProcessNewEntities()
+        {
+            Entities.With(newEntitiesQuery).ForEach(
+                (Entity entity, ref SpatialEntityId spatialEntityId, ref Metadata.Component metadata) =>
+                {
+                    var entityType = metadata.EntityType;
+                    var expectedTypes = entityTypeExpectations.GetExpectedTypes(entityType);
+
+                    foreach (var expectedType in expectedTypes)
+                    {
+                        if (!EntityManager.HasComponent(entity, expectedType))
+                        {
+                            return;
+                        }
+                    }
+
+                    try
+                    {
+                        gameObjectCreator.OnEntityCreated(entityType, new SpatialOSEntity(entity, EntityManager), Linker);
+                    }
+                    catch (Exception e)
+                    {
+                        Debug.LogException(e);
+                    }
+
+                    PostUpdateCommands.AddComponent(entity, new GameObjectInitSystemStateComponent
+                    {
+                        EntityId = spatialEntityId.EntityId
+                    });
+                });
         }
     }
 }